Rename - player model 위치 변경
Player model을 model 팩키지 아래에 player 팩키지를 생성하여 이동하였음.
